Shard the `convert.rs` module into sub-modules
Hopefully this'll make the organization a little nicer over time!
